+++ src/scene/shader-lib/chunks/gsplat/frag/gsplat.js
